refactor(#379): unify shared VPC network-adapter ip_address helpers
The OpenIaaS (#374) and VMware (#375) network-adapter resources each carried
their own copy of the VPC static-IP (ip_address) helpers; #375 had to use
vmware*-prefixed duplicates to avoid a symbol collision while both feature
branches were in flight. Both have now landed on rc/v1.9.0, so unify them.

- New internal/provider/vpc_network_adapter_ip.go holds the canonical shared
  logic: vpcStaticIPToPush, adapterVPCStaticIP, and a generic
  ensureVPCForIPAddress(ctx, d, networkVPCStatusFunc) whose only per-surface
  difference — the network read (OpenIaaS vs vCenter) — is injected as a
  callback (openIaasNetworkVPCBacked / vmwareNetworkVPCBacked).
- Removed the duplicated vmwareVPCStaticIPToPush / vmwareAdapterVPCStaticIP /
  ensureVMwareVPCForIPAddress and the now-redundant VMware unit test.
- Extracted validateIPAddressTargetsVPC (the pure fail-closed verdict) and
  unit-tested it directly, so the "reject ip_address on a non-VPC or absent
  network before any side effect" contract is now pinned for both resources at
  once, where it was previously covered only by live E2E.

Behaviour-preserving: schema golden snapshot and docs are unchanged. The only
user-visible change is the VMware validation error wording, aligned to the
OpenIaaS one ("not a VPC-backed private network").

Refs #379
