+++ internal/provider/resource_compute_iaas_opensource_network_adapter.go
@@ -135,7 +135,7 @@ func openIaasNetworkAdapterCreate(ctx context.Context, d *schema.ResourceData, m
 	vmID := d.Get("virtual_machine_id").(string)
 
 	// Reject ip_address on a non-VPC network BEFORE creating anything.
-	if diags := ensureVPCForIPAddress(ctx, c, d); diags != nil {
+	if diags := ensureVPCForIPAddress(ctx, d, openIaasNetworkVPCBacked(c)); diags != nil {
 		return diags
 	}
 
@@ -298,7 +298,7 @@ func openIaasNetworkAdapterUpdate(ctx context.Context, d *schema.ResourceData, m
 	c := getClient(meta)
 
 	// Reject ip_address on a non-VPC network BEFORE moving the adapter.
-	if diags := ensureVPCForIPAddress(ctx, c, d); diags != nil {
+	if diags := ensureVPCForIPAddress(ctx, d, openIaasNetworkVPCBacked(c)); diags != nil {
 		return diags
 	}
 
@@ -466,65 +466,21 @@ func openIaasNetworkAdapterDelete(ctx context.Context, d *schema.ResourceData, m
 	return nil
 }
 
-// ensureVPCForIPAddress fails closed BEFORE any side effect when ip_address is
-// explicitly configured but the target network_id is not a VPC-backed private
-// network. ip_address only has meaning on a VPC network (it assigns the
-// adapter's static IP there); on a plain network the platform ignores it, so
-// the value could never be applied nor recorded and the plan would never
-// converge. Validating the target network up front makes apply reject the bad
-// config instead of creating/moving the adapter and only then failing.
-func ensureVPCForIPAddress(ctx context.Context, c *client.Client, d *schema.ResourceData) diag.Diagnostics {
-	raw := d.GetRawConfig()
-	if raw.IsNull() {
-		return nil
+// openIaasNetworkVPCBacked reads the OpenIaaS network's VPC-ness for
+// ensureVPCForIPAddress (the decision logic is shared with the VMware adapter;
+// see vpc_network_adapter_ip.go). It is the only OpenIaaS-specific part of the
+// ip_address pre-validation: the network read.
+func openIaasNetworkVPCBacked(c *client.Client) networkVPCStatusFunc {
+	return func(ctx context.Context, networkID string) (vpcBacked bool, found bool, err error) {
+		network, err := c.Compute().OpenIaaS().Network().Read(ctx, networkID)
+		if err != nil {
+			return false, false, err
+		}
+		if network == nil {
+			return false, false, nil
+		}
+		return network.VPC != nil, true, nil
 	}
-	if v := raw.GetAttr("ip_address"); v.IsNull() {
-		return nil
-	}
-	ip := d.Get("ip_address").(string)
-	if ip == "" {
-		return nil
-	}
-	networkID := d.Get("network_id").(string)
-	network, err := c.Compute().OpenIaaS().Network().Read(ctx, networkID)
-	if err != nil {
-		return diag.Errorf("failed to read network %s to validate ip_address: %s", networkID, err)
-	}
-	if network == nil {
-		return diag.Errorf("network %s not found while validating ip_address", networkID)
-	}
-	if network.VPC == nil {
-		return diag.Errorf("ip_address %q is set but network %s is not a VPC-backed private network: ip_address requires network_id to reference a VPC private network — remove ip_address or use a VPC network", ip, networkID)
-	}
-	return nil
-}
-
-// vpcStaticIPToPush decides the VPC static IP to send on an OpenIaaS adapter
-// update; it returns the address to push, or "" to leave the static IP
-// untouched. It pushes ONLY when ip_address is explicitly configured, non-empty,
-// the adapter is on a VPC-backed network, and the configured address diverges
-// from the live one: re-sending an unchanged address would relocate the static
-// IP to itself on every apply (a perpetual no-op activity), and ip_address has
-// no meaning on a non-VPC network. The live static IP is resolved by MAC by the
-// caller, because the platform does not echo it on the adapter object (#1854).
-func vpcStaticIPToPush(ipConfigured bool, configuredIP, liveIP string, onVPC bool) string {
-	if !ipConfigured || configuredIP == "" || !onVPC {
-		return ""
-	}
-	if configuredIP == liveIP {
-		return ""
-	}
-	return configuredIP
-}
-
-// adapterVPCStaticIP maps a by-MAC static IP read to the ip_address state value.
-// A non-VPC adapter has no static IP; a VPC adapter with none registered yet
-// (sip == nil — including the 403/absent the client maps to nil) also yields "".
-func adapterVPCStaticIP(onVPC bool, sip *client.StaticIP) string {
-	if !onVPC || sip == nil {
-		return ""
-	}
-	return sip.IPAddress
 }
 
 // vifCleanupTargets partitions the adapter ids referenced by the failed
