+++ internal/provider/resource_compute_iaas_opensource_network_adapter_unit_test.go
@@ -65,47 +65,3 @@ func TestClassifyMissingDevice(t *testing.T) {
 		t.Fatalf("present tenant-wide only is drift (detached/moved), never a deletion, got %v", got)
 	}
 }
-
-// TestVPCStaticIPToPush pins the update decision for the VPC static IP
-// (ip_address, #374). The subtle rules: never push when unconfigured/empty/off-VPC,
-// never re-push an unchanged address (it would relocate the static IP to itself
-// on every apply), push on a genuine divergence or a first set.
-func TestVPCStaticIPToPush(t *testing.T) {
-	cases := []struct {
-		name                 string
-		ipConfigured         bool
-		configuredIP, liveIP string
-		onVPC                bool
-		want                 string
-	}{
-		{"not configured -> never push", false, "", "", true, ""},
-		{"configured but empty -> never push", true, "", "", true, ""},
-		{"configured non-empty but not on a VPC network -> never push", true, "192.168.0.10", "", false, ""},
-		{"configured, on VPC, no live IP yet -> push (first set)", true, "192.168.0.10", "", true, "192.168.0.10"},
-		{"configured equals live -> no redundant relocate-to-self", true, "192.168.0.10", "192.168.0.10", true, ""},
-		{"configured diverges from live -> relocate", true, "192.168.0.11", "192.168.0.10", true, "192.168.0.11"},
-		{"a stale configured value with ipConfigured=false (field cleared) -> never push", false, "192.168.0.11", "192.168.0.10", true, ""},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			if got := vpcStaticIPToPush(c.ipConfigured, c.configuredIP, c.liveIP, c.onVPC); got != c.want {
-				t.Fatalf("vpcStaticIPToPush(%v,%q,%q,%v) = %q, want %q", c.ipConfigured, c.configuredIP, c.liveIP, c.onVPC, got, c.want)
-			}
-		})
-	}
-}
-
-// TestAdapterVPCStaticIP pins the read mapping for ip_address (#374): a non-VPC
-// adapter and a VPC adapter without a registered static IP both yield empty; a
-// VPC adapter reflects the by-MAC static IP address.
-func TestAdapterVPCStaticIP(t *testing.T) {
-	if got := adapterVPCStaticIP(false, &client.StaticIP{IPAddress: "1.2.3.4"}); got != "" {
-		t.Fatalf("a non-VPC adapter has no static IP; want empty, got %q", got)
-	}
-	if got := adapterVPCStaticIP(true, nil); got != "" {
-		t.Fatalf("a VPC adapter with no registered static IP (nil) must yield empty, got %q", got)
-	}
-	if got := adapterVPCStaticIP(true, &client.StaticIP{IPAddress: "192.168.0.10"}); got != "192.168.0.10" {
-		t.Fatalf("a VPC adapter must reflect the by-MAC static IP; want 192.168.0.10, got %q", got)
-	}
-}
